x86-64 fixes

+++ code/qcommon/msg.c
@@ -796,7 +796,7 @@ void MSG_ReportChangeVectors_f( void ) {
 
 typedef struct {
 	char	*name;
-	int		offset;
+	size_t	offset;
 	int		bits;		// 0 = float
 } netField_t;
